fix build bugs

+++ cpp/lesscpp/build-linux.py
@@ -1,5 +1,6 @@
 import shutil,os
-os.remove('config.py')
+if os.path.exits('config.py'):
+	os.remove('config.py')
 shutil.copy('config-linux.py','config.py')
 os.system("scons -j6")
